# qb-network-graph-ui/docs/api-spec.yaml
openapi: 3.1.0
info:
  title: QB Network Graph API
  version: 1.0.0
  description: |
    Backend API for the QB Network Graph UI. Provides entity resolution,
    relationship traversal, native perspective overrides, and AI-assisted
    network analysis for the QuickBooks business network.
  contact:
    name: QB Network Graph Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development

# ─────────────────────────────────────────────────────────
# Tags
# ─────────────────────────────────────────────────────────
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────
# Component Schemas
# ─────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token injected by client auth interceptor (not yet implemented)
  schemas:
...
    Entity:
      type: object
      required: [id, name, industry, naics, city, state, confidence, vendors, clients, volume, variants, commodities]
      properties:
        id:
          type: string
          example: e1
        name:
          type: string
          example: Acme Construction Co
        ein:
          type: string
          nullable: true
          example: "74-3201587"
        contactName:
          type: string
          nullable: true
          example: John Mitchell
        email:
          type: string
          format: email
          nullable: true
          example: john@acmeconstruction.com
        phone:
          type: string
          nullable: true
          example: "(512) 555-0100"
        website:
          type: string
          format: uri
          nullable: true
          example: https://acmeconstruction.com
        industry:
          type: string
          description: NAICS code
          example: "236220"
        naics:
          type: string
          description: Same as industry (canonical NAICS code)
          example: "236220"
        legalStructure:
          type: string
          enum: [LLC, Corp, Inc, Sole Prop]
          example: LLC
        address:
          type: string
          nullable: true
          example: 4500 Congress Ave
        city:
          type: string
          example: Austin
        state:
          type: string
          example: TX
        zip:
          type: string
          nullable: true
          example: "78701"
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Entity resolution confidence score
          example: 0.96
        vendors:
          type: integer
          description: Count of vendor relationships
          example: 23
        clients:
          type: integer
          description: Count of client relationships
          example: 45
        volume:
          type: number
          description: Annual transaction volume in USD
          example: 4200000
        variants:
          type: array
          items:
            type: string
          description: Known name variants
          example: ["Acme Const", "ACME Construction"]
        commodities:
          type: array
          items:
            type: string
          description: Product/service categories
          example: ["Building Materials", "Concrete", "Steel"]
        serviceArea:
          type: string
          nullable: true
          description: Geographic service region
          example: Central Texas
        x:
          type: number
          description: Graph layout x-coordinate
          example: 400
        y:
          type: number
          description: Graph layout y-coordinate
          example: 300
    Relationship:
      type: object
      required: [source, target, volume, count, status]
      properties:
        source:
          type: string
          description: Payer entity ID
          example: e1
        target:
          type: string
          description: Payee entity ID
          example: e2
        volume:
          type: number
          description: Annual transaction volume in USD
          example: 450000
        count:
          type: integer
          description: Number of transactions
          example: 38
        status:
          type: string
          enum: [active, dormant]
          example: active
    MonthlyVolume:
      type: object
      required: [month, vol]
      properties:
        month:
          type: string
          description: Abbreviated month name
          example: Jan
        vol:
          type: number
          description: Volume value (in thousands USD)
          example: 480
...
    Error:
      type: object
      required: [message, status]
      properties:
        message:
          type: string
          example: Entity not found
        status:
          type: integer
          example: 404

# ─────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────
paths:
...
  /entities:
    get:
      operationId: getEntities
      tags: [Entities]
      summary: List all entities
      description: Returns all golden-record entities. Supports filtering by text query and industry code.
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Text search across entity name and variants
        - name: industry
          in: query
          schema:
            type: string
          description: Filter by NAICS industry code
      responses:
        "200":
          description: Entity list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Entity"
                  total:
                    type: integer
                    example: 10
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /entities/{id}:
    get:
      operationId: getEntity
      tags: [Entities]
      summary: Get a single entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: e1
      responses:
        "200":
          description: Entity detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Entity"
        "404":
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      operationId: updateEntity
      tags: [Entities]
      summary: Partially update an entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: e1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Any subset of Entity fields to update
      responses:
        "200":
          description: Updated entity
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Entity"
        "404":
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Relationships ────────────────────────────────────
...
  /entities/{id}/relationships:
    get:
      operationId: getEntityRelationships
      tags: [Relationships]
      summary: Get relationships for an entity
      description: Returns all edges where the entity is either source or target.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: e1
      responses:
        "200":
          description: Entity relationships
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Relationship"
  /entities/{id}/network:
    get:
      operationId: getEntityNetwork
      tags: [Relationships]
      summary: Get ego network subgraph
      description: |
        BFS traversal from the given entity up to `depth` hops.
        Returns all reachable entities and the relationships between them.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: e1
        - name: depth
          in: query
          schema:
            type: integer
            default: 2
            minimum: 1
            maximum: 5
          description: Number of hops to traverse
      responses:
        "200":
          description: Network subgraph
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      entities:
                        type: array
                        items:
                          $ref: "#/components/schemas/Entity"
                      relationships:
                        type: array
                        items:
                          $ref: "#/components/schemas/Relationship"
  /entities/{id}/volume:
    get:
      operationId: getEntityVolume
      tags: [Relationships]
      summary: Get monthly transaction volume
      description: Returns time-series transaction volume data for charting.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: e1
      responses:
        "200":
          description: Monthly volume series
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MonthlyVolume"

  # ── Search ───────────────────────────────────────────
...